connect it to Contentful
