+++ src/shared/components/Contentful/MemberTalkCloud/themes/default.scss
@@ -27,7 +27,7 @@
 
 .entry {
   transition: all 0.4s;
-  margin: 20px;
+  margin: 40px 20px;
 
   img {
     cursor: pointer;
